cgroup: introduce proper cgroup/v1, cgroup/v2 packages
Currently we have func .*V[1|2].* named function names in the cgroup
package.

This factors those functions into two separate packages cgroup/v1, and
cgroup/v2.

This is a prerequisite for #3242
